Changes int64 variables to int

+++ utils/math.go
@@ -1,6 +1,6 @@
 package utils
 
-func DivMod(numerator, denominator int64) (quotient, remainder int64) {
+func DivMod(numerator, denominator int) (quotient, remainder int) {
 	quotient = numerator / denominator // integer division, decimals are truncated
 	remainder = numerator % denominator
 	return
